:recycle: Resctructured files, organized into classes
Everything should be a class--still a WIP. Changing the play.py file to just include a Game object and run the game. All logic for the game,round, or trick will be handled in the correct scope. Once the logic is figured out, we will move on to makeing it send via sockets.